Pretreat/requirement2024.txt
absl-py==0.12.0
affine==2.3.0
aiohttp==3.6.2
alabaster==0.7.12
amqp==5.1.1
ansi2html==1.8.0
appnope==0.1.0
arcade==2.2.5
argcomplete==1.11.1
argh==0.26.2
arghandler==1.2.0
args==0.1.0
arrow==0.12.1
asgiref==3.3.0
astor==0.8.0
astroid==2.2.5
astunparse==1.6.3
async-timeout==4.0.2
atari-py==0.2.9
attrs==19.1.0
autograd==1.6.2
autopep8==1.5
av==8.0.2
Babel==2.7.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bandit==1.6.0
billiard==3.6.4.0
biopython==1.79
bleach==3.1.0
boto==2.49.0
Box2D==2.3.10
box2d-py==2.3.8
braceexpand==0.1.5
Brotli==1.0.9
c==0.1.0
cached-property==1.5.2
cachetools==4.0.0
cdd==0.1.4
celery==5.1.2
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.10
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
client==0.0.1
cligj==0.7.2
clint==0.5.1
cloudpickle==1.2.2
cmake==3.18.4.post1
coco==0.4.0
colorama==0.4.1
colour==0.1.5
contextily==1.1.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.13
dash==1.18.1
dash-bio==1.0.2
dash-bootstrap-components==0.11.0
dash-core-components==1.14.1
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dataclasses==0.7
dateparser==1.1.3
dbr==8.1.2
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.14
dill==0.3.4
diskcache==5.6.3
Django==3.1.3
djitellopy==2.4.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.15.2
dronekit==2.9.2
dronekit-sitl==3.3.0
entrypoints==0.3
enums==0.0.2
et-xmlfile==1.0.1
ffmpeg==1.4
filterpy==1.4.5
Fiona==1.8.20
Flask==1.1.1
Flask-Compress==1.8.0
flatbuffers==1.12
frozendict==2.3.0
frozenlist==1.2.0
fuel==0.2.0
future==0.18.2
gast==0.3.3
geographiclib==1.52
Geometry==0.0.23
geopandas==0.9.0
GEOparse==2.0.3
geopy==2.2.0
geos==0.2.3
gitdb2==2.0.5
GitPython==2.1.11
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gurobipy==9.1.2
gym==0.18.3
h5py==2.10.0
haversine==2.1.2
httpie==1.0.2
idna==2.8
idna-ssl==1.1.0
igraph==0.9.8
image==1.5.33
imageai==2.1.5
imageio==2.5.0
imagesize==1.1.0
imbalanced-learn==0.8.1
imblearn==0.0
imgaug==0.3.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.3
install==1.3.4
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==1.0.2
isort==4.3.20
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
joblib==1.0.1
json5==0.8.5
jsonpickle==2.0.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-dash==0.4.2
jupyterlab-server==1.0.6
kaleido==0.2.1
Keras==2.4.3
Keras-Applications==1.0.8
keras-core==0.0.0
Keras-Preprocessing==1.1.2
keras-rl2==1.0.4
keras-tuner==1.4.4
keyring==8.7
keyrings.alt==1.3
kiwisolver==1.3.1
kombu==5.1.0
kt-legacy==1.0.5
lazy-object-proxy==1.4.1
llvmlite==0.36.0
lxml==4.4.2
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.3.4
mccabe==0.6.1
mercantile==1.2.1
mistune==0.8.4
mock==3.0.5
mongotriggers==2.0.1
monotonic==1.5
more-itertools==8.1.0
mpld3==0.5.7
mpmath==1.1.0
multidict==4.5.2
munch==2.5.0
mydict==1.0.20
natsort==8.1.0
nbconvert==5.5.0
nbformat==4.4.0
nest-asyncio==1.5.6
networkx==2.3
nltk==3.4.1
nose==1.3.7
notebook==5.7.8
numba==0.53.1
numexpr==2.7.1
numpy==1.19.5
oauthlib==3.1.0
object-detection==0.1
odfpy==1.4.0
odict==1.7.0
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.1.1.26
openpyxl==3.0.1
opt-einsum==3.3.0
ordered-enum==0.0.6
packaging==21.3
pandas==1.1.5
pandocfilters==1.4.2
ParmEd==3.4.3
parso==0.4.0
pathtools==0.1.2
patsy==0.5.3
pbr==5.2.1
periodictable==1.6.1
pexpect==4.7.0
picklable-itertools==0.1.1
pickle5==0.0.12
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.1.2
pipenv==2018.11.26
plat==0.2.2
plotly==4.14.1
pluggy==0.13.1
plumber==1.5
progressbar2==3.50.1
prometheus-client==0.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.0
pycodestyle==2.5.0
pydot==1.4.2
pygame==2.1.0
pyglet==1.5.11
pyglet-ffmpeg2==0.1.16
Pygments==2.4.2
pyicloud-ipd==0.10.1
pyknow==1.7.0
pylint==2.3.1
pymavlink==2.4.0
pymongo==4.0.1
pynndescent==0.5.10
pyny3d==0.2
pyparsing==2.4.7
pyproj==3.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.15.2
pyserial==3.5
pyshp==2.1.0
pytest==5.3.4
python-crontab==2.4.0
python-dateutil==2.8.1
python-utils==2.4.0
pytiled-parser==0.9.2
pytz==2019.1
pytz-deprecation-shim==0.1.0.post0
pyvis==0.1.9
pyvisgraph==0.2.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
rasterio==1.2.10
redis==4.3.6
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.0
retrying==1.3.3
rr.opt.mcts.simple==0.3.0
rsa==4.0
schema==0.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scikit-lego==0.6.14
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.8.0
sip==5.1.0
six==1.15.0
sklearn==0.0
sklearn-pandas==2.0.4
sklearn2pmml==0.98.1
sklego==0.0
smmap2==2.0.5
snowballstemmer==2.0.0
snuggs==1.4.7
sortedcontainers==2.1.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.4.1
statsmodels==0.12.2
stevedore==1.30.1
svn==1.0.1
tables==3.6.1
tablib==0.14.0
tabulate==0.8.2
tbb==2021.10.0
tello-python==1.0.5
tellopy==0.6.0
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==1.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==1.6.4
tf==1.0.0
Theano==1.0.5
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.0
tornado==6.0.3
tqdm==4.56.0
traitlets==4.3.2
typed-ast==1.3.5
typing-extensions==3.7.4
tzdata==2022.1
tzlocal==4.2
umap-learn==0.5.4
urllib3==1.25.6
utils==1.0.1
utm==0.4.2
vine==5.0.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wolframclient==1.1.0
wolframwebengine==1.0.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.3.0
z3==0.2.0
z3-solver==4.8.10.0
zipp==3.6.0
zmq==0.0.0
zope.component==4.6
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
zope.lifecycleevent==4.3
zope.proxy==4.3.3
